+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This file only reflects the changes that are made in this image.
 Please refer to the [Salt 3005 Release Notes](https://docs.saltstack.com/en/latest/topics/releases/3005.html)
 for the list of changes in SaltStack.
 
+**3005_1**
+
+- Add support for setting the `salt-api` user's password via docker secrets using the environment
+  variable `SALT_API_USER_PASS_FILE`. Has priority over `SALT_API_USER_PASS`.
+
 **3005**
 
 - Upgrade `salt-master` to `3005` *Phosphorus*
@@ -111,7 +116,8 @@ Support for the `USERMAP_UID` and `USERMAP_GID` env variables will be removed wi
 
 - Upgrade `salt-master` to `3002.1` *Magnesium*
 - Change Docker base image to `ubuntu:focal-20201008`
-- Fix issue changing _read-only_ directories ownership ([@Kidswiss](https://github.com/Kidswiss) - [#47](https://github.com/cdalvaro/docker-salt-master/pull/47))
+- Fix issue changing _read-only_ directories ownership
+  ([@Kidswiss](https://github.com/Kidswiss) - [#47](https://github.com/cdalvaro/docker-salt-master/pull/47))
 
 **3002**